+++ src/components/landing/B2B.jsx
@@ -1,5 +1,6 @@
 import { Conclusion, Heading } from ".";
 import "./landing.css";
+import RightArrow from "../../assets/icon/right-arrow.svg";
 import NefentusLogo from "../../assets/logo/logo.svg";
 import HeroLineTop from "../../assets/landing/b2c-hero-top.svg";
 import HeroLineBottom from "../../assets/landing/b2c-hero-down.svg";
@@ -30,61 +31,19 @@ import { ScrollTrigger } from "gsap/ScrollTrigger";
 gsap.registerPlugin(ScrollTrigger);
 
 const B2B = () => {
-  const [step, setStep] = useState(0);
-  const [progress, setProgress] = useState(0);
-  const [isPositionedAtStart, setIsPositionedAtStart] = useState(false);
-
   const b2bContainer = useRef();
 
-  const handleScroll = (e) => {
-    if (document.documentElement.scrollTop === 0) {
-      setIsPositionedAtStart(true);
-    }
-  };
-
   useEffect(() => {
-    // Ensure the component has mounted in a browser environment
-    window.addEventListener("scroll", handleScroll);
-
-    // Scroll to the top when component mounts
-    setIsPositionedAtStart(document.documentElement.scrollTop === 0);
-    return () => {
-      window.removeEventListener("scroll", handleScroll);
-    };
+    const href = window.location.href.substring(
+      window.location.href.lastIndexOf("#") + 1,
+    );
+    setTimeout(() => {
+      const element = document.getElementById(href);
+      if (element) {
+        element.scrollIntoView();
+      }
+    }, 500);
   }, []);
-
-  useEffect(() => {
-    if (isPositionedAtStart) {
-      // Initialize GSAP
-      const numImages = products.length;
-
-      const tl = gsap.from(".b2b .product", {
-        scrollTrigger: {
-          trigger: ".layout-paragraph.product .layout-title .sub-title",
-          start: "top top",
-          end: () => `+=${numImages * 80}%`,
-          scrub: true,
-          pin: ".landing-layout.container.b2b",
-          onUpdate: (self) => {
-            setStep(Math.floor(self.progress * numImages));
-            setProgress(Math.floor(self.progress * numImages * 100));
-          },
-          onLeave: () => {
-            setStep(numImages - 1);
-          },
-          onLeaveBack: () => {
-            setStep(0);
-          },
-        },
-        duration: 1,
-        ease: "none",
-      });
-
-      return () => {
-        tl.kill(); // Ensure GSAP instance is killed
-      };
-    }
-  }, [isPositionedAtStart]);
 
   return (
     <div>
@@ -99,7 +58,7 @@ const B2B = () => {
 
         <Invoicing />
 
-        <Product stepId={step} progress={progress} />
+        <Product />
 
         <Safe />
 
@@ -124,6 +83,7 @@ export default B2B;
 const Hero = () => {
   return (
     <div
+      id="hero"
       className="layout-paragraph hero"
       style={{
         paddingBlock: "0",
@@ -148,9 +108,15 @@ const Hero = () => {
             Seamlessly create & receive payments. Manage all your wallets and
             streamline your sales process—all from a single platform.
           </p>
-          <a href="/signup">
-            <button>Get Started</button>
-          </a>
+          <div className="button-container">
+            <a href="/signup">
+              <button>Get Started</button>
+            </a>
+            <a href="/contact">
+              <span>Talk to an expert</span>
+              <img src={RightArrow} />
+            </a>
+          </div>
         </div>
         <img
           src={HeroPng}
@@ -278,8 +244,9 @@ const Invoicing = () => {
       clearInterval(timeRef.current);
     };
   }, []);
+
   return (
-    <div className="layout-paragraph">
+    <div id="invoicing" className="layout-paragraph">
       <Heading
         title={`Effortless invoice creation`}
         subtitle={`Use our built-in wallet or link an unlimited number of external wallets for seamless transactions.`}
@@ -310,6 +277,7 @@ const Invoicing = () => {
                       return false;
                     }),
                   );
+                clearInterval(timeRef.current);
               }}
             >
               <img src={item.icon} />
@@ -334,6 +302,7 @@ const Invoicing = () => {
                       return false;
                     }),
                   );
+                clearInterval(timeRef.current);
               }}
             >
               <img src={item.icon} />
@@ -380,8 +349,31 @@ const products = [
     img: Product3Png,
   },
 ];
-const Product = ({ stepId, progress }) => {
+const Product = () => {
+  const [stepId, setStepId] = useState(0);
+  const [progress, setProgress] = useState(0);
   const [step, setStep] = useState(products.map((item, id) => !id));
+  const timeRef = useRef();
+  const updateProgress = () => {
+    if (progress <= 300) {
+      const newProgress = progress + 0.3;
+      setProgress(newProgress);
+      setStepId(Math.trunc(progress / 100));
+    } else {
+      setProgress(0);
+      setStepId(0);
+    }
+  };
+  const start = () => {
+    clearInterval(timeRef.current);
+    timeRef.current = setInterval(updateProgress, 10);
+  };
+  useEffect(() => {
+    start();
+    return () => {
+      clearInterval(timeRef.current);
+    };
+  }, [progress]);
 
   useEffect(() => {
     setStep((prev) => {
@@ -392,7 +384,7 @@ const Product = ({ stepId, progress }) => {
     });
   }, [stepId]);
   return (
-    <div className="layout-paragraph product">
+    <div id="product" className="layout-paragraph product">
       <Heading
         title={`Simplified <span style="text-wrap: nowrap">product creation</span>`}
         subtitle={`Utilize our internal wallet or link an unlimited number of external wallets for hassle-free product creation.`}
@@ -408,14 +400,16 @@ const Product = ({ stepId, progress }) => {
             <div
               className={`product-step`}
               onClick={() => {
-                !step[id] &&
-                  id !== undefined &&
+                if (!step[id] && id !== undefined) {
                   setStep(
                     step.map((s, i) => {
                       if (i == id) return !s;
                       return false;
                     }),
                   );
+                  setStepId(id);
+                  setProgress(id * 100);
+                }
               }}
               style={{
                 width: `calc(${100 / products.length}%)`,
@@ -615,6 +609,7 @@ const Safe = () => {
 const Analytics = () => {
   return (
     <div
+      id="sales"
       className="layout-paragraph analytics"
       style={{
         paddingBottom: "0",