+++ src/components/navigation/navigation.jsx
@@ -9,6 +9,25 @@ import { logOut } from "../../utils";
 import UserProfile from "../userProfile/userProfile";
 import { useTheme } from "../../context/themeContext/themeContext";
 import Cookie from "js-cookie";
+import MultiWallet from "../../assets/icon/wallet-01.svg";
+import MultiWalletBlue from "../../assets/icon/wallet-01 - blue.svg";
+import InternalWallet from "../../assets/icon/wallet-02.svg";
+import InternalWalletBlue from "../../assets/icon/wallet-02 - blue.svg";
+import Sales from "../../assets/icon/pie-chart-01.svg";
+import SalesBlue from "../../assets/icon/pie-chart-01 - blue.svg";
+import Invoicing from "../../assets/icon/receipt-check.svg";
+import InvoicingBlue from "../../assets/icon/receipt-check - blue.svg";
+import Shopping from "../../assets/icon/shopping-bag-01.svg";
+import ShoppingBlue from "../../assets/icon/shopping-bag-01 - blue.svg";
+import Convert from "../../assets/landing/swap-ico.svg";
+import ConvertBlue from "../../assets/landing/swap-ico - blue.svg";
+import ContactSupport from "../../assets/icon/contact-support.svg";
+import ContactSupportBlue from "../../assets/icon/contact-support-blue.svg";
+import ContactExpert from "../../assets/icon/contact-expert.svg";
+import ContactExpertBlue from "../../assets/icon/contact-expert-blue.svg";
+import Send from "../../assets/landing/send-ico.svg";
+import Dropdown from "../../assets/icon/dropdown.svg";
+import { hover } from "@testing-library/user-event/dist/hover";
 
 const Navigation = () => {
   const { theme, toggleTheme } = useTheme();
@@ -128,28 +147,23 @@ const Navigation = () => {
             {!hideOptions && (
               <ul className={styles.navList}>
                 <li className="standard">
-                  <Link to="/b2c">
-                    <p>Personal</p>
-                    <p className={styles.fake}>Personal</p>
-                  </Link>
+                  <Products />
                 </li>
                 <li className="standard">
-                  <Link to="/b2b">
-                    <p>Business Solutions</p>
-                    <p className={styles.fake}>Business Solutions</p>
-                  </Link>
-                </li>
-                <li className="standard">
-                  <Link to="/resources">
-                    <p>Resources</p>
-                    <p className={styles.fake}>Resources</p>
-                  </Link>
+                  <Resources />
                 </li>
                 <li className="standard">
                   <Link to="/vacancy">
                     <p>Career</p>
-                    <p className={styles.fake}>Career</p>
                   </Link>
+                </li>
+                {/* <li className="standard">
+                  <Link to="/contact">
+                    <p>Contact us</p>
+                  </Link>
+                </li> */}
+                <li className="standard">
+                  <ContactUs />
                 </li>
               </ul>
             )}
@@ -202,6 +216,7 @@ const Navigation = () => {
       </div>
 
       <div
+        id="mobileMenu"
         className={`${styles.mobileMenu}`}
         style={{
           transform: openMenu ? "translateY(0%)" : "translateY(-120%)",
@@ -209,22 +224,21 @@ const Navigation = () => {
       >
         <div>
           <ul>
-            <Link to="/b2c" onClick={() => setOpenMenu(false)}>
-              <li className="standard">Personal</li>
-            </Link>
-            <Link to="/b2b" onClick={() => setOpenMenu(false)}>
-              <li className="standard">Business Solutions</li>
-            </Link>
-
-            {/* <Link to="/affiliate" onClick={() => setOpenMenu(false)}>
-              <li className="standard">{t("navigation.affiliate")}</li>
-            </Link> */}
-            <Link to="/resources" onClick={() => setOpenMenu(false)}>
-              <li className="standard">Resources</li>
-            </Link>
+            <li>
+              <Products setOpenMenu={setOpenMenu} />
+            </li>
+            <li>
+              <Resources />
+            </li>
             <Link to="/vacancy" onClick={() => setOpenMenu(false)}>
               <li className="standard">Career</li>
             </Link>
+            {/* <Link to="/contact" onClick={() => setOpenMenu(false)}>
+              <li className="standard">Contact us</li>
+            </Link> */}
+            <li>
+              <ContactUs />
+            </li>
             {loginAndSignupMobile()}
           </ul>
         </div>
@@ -234,3 +248,258 @@ const Navigation = () => {
 };
 
 export default Navigation;
+
+const DropDownMenuItem = ({ icon, title, subtitle }) => {
+  const [ico, setIcon] = useState(icon.normal);
+  return (
+    <div
+      onMouseEnter={() => setIcon(icon.hover)}
+      onMouseLeave={() => setIcon(icon.normal)}
+      className={styles.dropdownMenuItem}
+    >
+      <div className={styles.iconWrapper}>
+        <img src={ico} />
+      </div>
+      <div className={styles.body}>
+        <p>{title}</p>
+        <p>{subtitle}</p>
+      </div>
+    </div>
+  );
+};
+
+const Products = ({ setOpenMenu }) => {
+  const [open, setOpen] = useState(false);
+
+  return (
+    <div
+      style={{
+        position: "relative",
+      }}
+      onMouseEnter={() => setOpen(true)}
+      onMouseLeave={() => setOpen(false)}
+      // onTouchEnd={(e) => setOpen((prev) => !prev) && e.preventDefault()}
+    >
+      <div className={styles.products}>
+        <p>Products</p>
+        <img
+          src={Dropdown}
+          style={{
+            transition: "0.2s",
+            transform: open ? "rotate(180deg)" : "",
+          }}
+        />
+      </div>
+      {open && (
+        <div className={styles.dropdown}>
+          <div
+            style={{
+              padding: "1.4rem",
+              display: "flex",
+              flexDirection: "column",
+              borderRight: "1px solid #323232",
+            }}
+          >
+            <a href="/b2b" onClick={() => setOpenMenu(false)}>
+              <p
+                style={{
+                  color: "#b1b1b1",
+                  padding: "0.8rem",
+                }}
+              >
+                Business tools
+              </p>
+            </a>
+            <a href="/b2b#hero" onClick={() => setOpenMenu(false)}>
+              <DropDownMenuItem
+                icon={{ normal: Shopping, hover: ShoppingBlue }}
+                title={`Cryptocurrency payments`}
+                subtitle={`Accept cryptocurrencies using unique payment links`}
+              />
+            </a>
+            <a href="/b2b#invoicing" onClick={() => setOpenMenu(false)}>
+              <DropDownMenuItem
+                icon={{ normal: Invoicing, hover: InvoicingBlue }}
+                title={`Invoicing`}
+                subtitle={`Create and send professional invoices that can be paid in cryptocurrency`}
+              />
+            </a>
+            <a href="/b2b#sales" onClick={() => setOpenMenu(false)}>
+              <DropDownMenuItem
+                icon={{ normal: Sales, hover: SalesBlue }}
+                title={`Sales dashboard`}
+                subtitle={`Track your sales and monitor transaction trends`}
+              />
+            </a>
+          </div>
+          <div
+            style={{
+              padding: "1.4rem",
+              display: "flex",
+              flexDirection: "column",
+            }}
+          >
+            <a href="/b2c" onClick={() => setOpenMenu(false)}>
+              <p
+                style={{
+                  color: "#b1b1b1",
+                  padding: "0.8rem",
+                }}
+              >
+                Personal use
+              </p>
+            </a>
+            <a href="/b2c#multi-wallets" onClick={() => setOpenMenu(false)}>
+              <DropDownMenuItem
+                icon={{ normal: MultiWallet, hover: MultiWalletBlue }}
+                title={`Multi-wallet management`}
+                subtitle={`Manage multiple cryptocurrency wallets effortlessly in one place`}
+              />
+            </a>
+            <a href="/b2c#transactions" onClick={() => setOpenMenu(false)}>
+              <DropDownMenuItem
+                icon={{ normal: Convert, hover: ConvertBlue }}
+                title={`Converter`}
+                subtitle={`Instantly convert between different cryptocurrencies with ease`}
+              />
+            </a>
+            <a href="/b2c#multi-wallets">
+              <DropDownMenuItem
+                icon={{ normal: InternalWallet, hover: InternalWalletBlue }}
+                title={`Nefentus wallet`}
+                subtitle={`Manage your digital assets securely within the Nefentus ecosystem`}
+              />
+            </a>
+          </div>
+        </div>
+      )}
+    </div>
+  );
+};
+
+const Resources = () => {
+  const [open, setOpen] = useState(false);
+
+  return (
+    <div
+      style={{ position: "relative" }}
+      onMouseEnter={() => setOpen(true)}
+      onMouseLeave={() => setOpen(false)}
+      onTouchEnd={(e) => setOpen((prev) => !prev) && e.preventDefault()}
+    >
+      <div className={styles.resources}>
+        <p>Resources</p>
+        <img
+          src={Dropdown}
+          style={{
+            transition: "0.2s",
+            transform: open ? "rotate(180deg)" : "",
+          }}
+        />
+      </div>
+      {open && (
+        <div className={styles.dropdown} style={{ display: "block" }}>
+          <div
+            style={{
+              padding: "1.4rem",
+              borderBottom: "1px solid #323232",
+            }}
+          >
+            <p
+              style={{
+                color: "#b1b1b1",
+                padding: "0.8rem",
+              }}
+            >
+              Docs
+            </p>
+            <div className={styles.gridContent}>
+              <a href="/resources">Guides and Tutorials</a>
+              {/* <a href="/resources">Case Studies and Testimonials</a>
+              <a href="/resources">API Documentation</a>
+              <a href="/resources">Blog and News</a>
+              <a href="/resources">Security Information</a> */}
+            </div>
+          </div>
+          <div
+            style={{
+              padding: "1.4rem",
+            }}
+          >
+            <p
+              style={{
+                color: "#b1b1b1",
+                padding: "0.8rem",
+              }}
+            >
+              Community
+            </p>
+            <div className={styles.gridContent}>
+              <a href="https://t.me/nefentus">
+                Telegram <img src={Send} style={{ width: "0.8rem" }} />
+              </a>
+              <a href="https://cy.linkedin.com/company/nefentuspay">
+                Linkedin <img src={Send} style={{ width: "0.8rem" }} />
+              </a>
+              <a href="https://x.com/nefentusapp">
+                Twitter <img src={Send} style={{ width: "0.8rem" }} />
+              </a>
+            </div>
+          </div>
+        </div>
+      )}
+    </div>
+  );
+};
+
+const ContactUs = () => {
+  const [open, setOpen] = useState(false);
+  return (
+    <div
+      style={{
+        position: "relative",
+      }}
+      onMouseEnter={() => setOpen(true)}
+      onMouseLeave={() => setOpen(false)}
+      onTouchEnd={(e) => setOpen((prev) => !prev) && e.preventDefault()}
+    >
+      <div className={styles.contactus}>
+        <p>Contact us</p>
+        <img
+          src={Dropdown}
+          style={{
+            transition: "0.2s",
+            transform: open ? "rotate(180deg)" : "",
+          }}
+        />
+      </div>
+      {open && (
+        <div className={styles.dropdown}>
+          <div
+            style={{
+              padding: "1.4rem",
+              display: "flex",
+              flexDirection: "column",
+              borderRight: "1px solid #323232",
+            }}
+          >
+            <a href="/technical-support" onClick={() => setOpenMenu(false)}>
+              <DropDownMenuItem
+                icon={{ normal: ContactSupport, hover: ContactSupportBlue }}
+                title={`Technical support`}
+                subtitle={`Report technical issues with the platform`}
+              />
+            </a>
+            <a href="/business-support" onClick={() => setOpenMenu(false)}>
+              <DropDownMenuItem
+                icon={{ normal: ContactExpert, hover: ContactExpertBlue }}
+                title={`Business support`}
+                subtitle={`Let us discuss how to integrate Nefentus into your own business`}
+              />
+            </a>
+          </div>
+        </div>
+      )}
+    </div>
+  );
+};