+++ CHANGELOG.md
@@ -24,6 +24,8 @@
 - Login Page right side image changed (MINOR)
 - Fix the responsiveness issue in setting page and add side bar in tablet and mobile view (PATCH)
 - fix the right users edit issue in users table (PATCH)
+- Fix landings && add contact page && update navbar (MINOR)
+- Fix Footer link Remove && Pay button in invoice preview && sidebar behavior (MiNOR)
 
 # 1.3.0